Serve development assets directly
Add a Rack::Static handler for the Webpack temporary location if
it exists, and serve the static assets from there.

Update Webpack and the Asset helper to not use a remote asset URL.

Update Procfile.dev to run the new watch yarn command, which
only compiles assets and doesn't serve them.